% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRegimenStatisticsTable.R
\name{createRegimenStats}
\alias{createRegimenStats}
\title{Create regimen stats table}
\usage{
createRegimenStats(
  connection,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  regimenStatsTable,
  regimenIngredientsTable,
  gapBetweenTreatment = 120
)
}
\arguments{
\item{regimenStatsTable}{A table what will be created in cohortDatabaseSchema}

\item{gapBetweenTreatment}{To calculate time to treatment discontinuation}
}
\value{
A data frame with cohort characteristics.
}
\description{
Computes features using all drugs, conditions, procedures, etc. observed on or prior to the cohort
index date.
}
