% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{getCohortCharacteristics}
\alias{getCohortCharacteristics}
\title{Create characterization of a cohort}
\usage{
getCohortCharacteristics(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  cohortId,
 
    covariateSettings = FeatureExtraction::createDefaultCovariateSettings(includedCovariateConceptIds
    = getCovariatesToInclude(), addDescendantsToInclude = TRUE,
    excludedCovariateConceptIds = 4162276, addDescendantsToExclude = TRUE)
)
}
\arguments{
\item{cohortId}{The cohort definition ID used to reference the cohort in the cohort
= table.}

\item{covariateSettings}{Either an object of type \code{covariateSettings} as created using one of
the createCovariate functions in the FeatureExtraction package, or a list
of such objects.}
}
\value{
A data frame with cohort characteristics.
}
\description{
Computes features using all drugs, conditions, procedures, etc. observed on or prior to the cohort
index date.
}
