% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{instantiateCohortSet}
\alias{instantiateCohortSet}
\title{Instantiate a set of cohort}
\usage{
instantiateCohortSet(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  cohortIds = NULL,
  createCohortTable = FALSE
)
}
\arguments{
\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the
construction to.}

\item{createCohortTable}{Create the cohort table? been executed.}

\item{generateInclusionStats}{Compute and store inclusion rule statistics?}

\item{inclusionStatisticsFolder}{The folder where the inclusion rule statistics are stored. Can be
left NULL if \code{generateInclusionStats = FALSE}.}
}
\description{
This function instantiates a set of cohort in the cohort table, using definitions that are fetched from a WebApi interface.
Optionally, the inclusion rule statistics are computed and stored in the \code{inclusionStatisticsFolder}.
}
