% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRegimenStatisticsTable.R
\name{createRegimenStats}
\alias{createRegimenStats}
\title{Create regimen stats table}
\usage{
createRegimenStats(
  connectionDetails,
  cdmDatabaseSchema,
  writeDatabaseSchema,
  cohortTable,
  regimenStatsTable,
  regimenIngredientsTable,
  gapBetweenTreatment = 120
)
}
\value{
The function returns nothing. As side effect it creates regimen stats table in write database schema
}
\description{
Create regimen stats table
}
